Move Console methods to separate extension class. Issue #60
