+++ DICOM/Log/Extensions.cs
@@ -1,12 +1,11 @@
 ﻿// Copyright (c) 2012-2015 fo-dicom contributors.
 // Licensed under the Microsoft Public License (MS-PL).
 
-using System;
-using System.Collections.Generic;
-using System.Text;
-
 namespace Dicom.Log
 {
+    using System.Collections.Generic;
+    using System.Text;
+
     public static class Extensions
     {
         public static void WriteToLog(this IEnumerable<DicomItem> dataset, Logger log, LogLevel level)
@@ -20,23 +19,6 @@ namespace Dicom.Log
             var logger = new DicomDatasetLogger(log, level);
             new DicomDatasetWalker(file.FileMetaInfo).Walk(logger);
             new DicomDatasetWalker(file.Dataset).Walk(logger);
-        }
-
-        public static void WriteToConsole(this IEnumerable<DicomItem> dataset)
-        {
-            var log = new StringBuilder();
-            var dumper = new DicomDatasetDumper(log, 80, 60);
-            new DicomDatasetWalker(dataset).Walk(dumper);
-            Console.WriteLine(log);
-        }
-
-        public static void WriteToConsole(this DicomFile file)
-        {
-            var log = new StringBuilder();
-            var dumper = new DicomDatasetDumper(log, 80, 60);
-            new DicomDatasetWalker(file.FileMetaInfo).Walk(dumper);
-            new DicomDatasetWalker(file.Dataset).Walk(dumper);
-            Console.WriteLine(log);
         }
 
         public static string WriteToString(this IEnumerable<DicomItem> dataset)
